+++ src/main/java/kr/co/pokemon/controller/MainController.java
@@ -17,39 +17,5 @@ public class MainController {
 //		model.addAttribute("pdto",playerDto);
 //		return "index";
 //	}
-	@GetMapping(value = "/friend/flist")
-	public String flist() {
-		return "/friend/flist";
-	}
 
-	
-	@GetMapping("/play/plist")
-	public String plist() {
-		return "/play/plist";
-	}	
-	
-	@GetMapping("/play/quest")
-	public String quest() {
-		return "/play/quest";
-	}
-	
-	@GetMapping("/play/dungeon")
-	public String dungeon() {
-		return "/play/dungeon";
-	}
-
-	@GetMapping("/play/sdungeon")
-	public String sdungeon() {
-		return "/play/sdungeon";
-	}
-	
-	@GetMapping("/first/first")
-	public String first() {
-		return "/first/first";
-	}	
-	
-	@GetMapping("/play/battle")
-	public String battle() {
-		return "/play/battle";
-	}
 }
