+++ src/main/java/kr/co/pokemon/plan/service/SdungeonService.java
@@ -1,0 +1,5 @@
+package kr.co.pokemon.plan.service;
+
+public interface SdungeonService {
+
+}
