+++ src/main/java/kr/co/pokemon/plan/service/SdungeonServiceImpl.java
@@ -1,0 +1,5 @@
+package kr.co.pokemon.plan.service;
+
+public class SdungeonServiceImpl implements SdungeonService {
+
+}
